Finished Tables tex doc & tidied up do files
(1) To facilitate the construction of the tex tables, I wrote duplicated sections of the esttab stata code,  in each do file, and removed the fragmented command.
(2) I used the tex tables output to created Full.tex to run the unfragmented tables, identify errors and edit the do files.
(3) I created the tables.tex file to compile all of the fragmented tex files.
(4) I deleted all of the unnecesary code in the do files and deleted the Full.tex file.

(5) NOTES: (A) the full_tables OG file is virtually unreadable; we should use the tables.tex file from now on.
(B) Table 10 does not automatically compile. This is because the "\$" in the model titles is read as a $ by latex. To force compile, just edit the doc after reunning project.