Summary Table 1 Complete
(1) Summary table 1: Listing Characteristics complete.
(1a) Melody wrote the code for the numerical variables, I wrote the code for the categorical variables.
(1b) NB I was wondering whether "Year of First Review" is correct or is actually what we would like in the final paper. Would 2014.93, for example, not be better?
(1c) NB This table was written fully in Stata, caption and notes included. I tried to include only the core of the table, and add thing in latex afterwards, but this proved problematic. Its probably doable, but I got a lot of glitches when I did it.
(2) Host Demographics table was started but not finished.
(3) document.tex was created so that we can run the talbes without having to run the entire project. This is a temporary file

TO DOs:
Finish host demographis table
Review Table 1