Table 2 is done!
I went over the issues I found yesturday. Here is a quick briefing:
(1) When we run the initial run_project, the total amount of observations is ~46k, not 69k like it is in your paper. THIS MAY BE A BIG ISSUE.
(2) All values sum up to correct amounts within ~regression output~. :)
(3) I will need your input interms of  formating 0s and 1s.