Full Data sample size
TO-DO:
(1) Figure out why the Full Data values for Property Types and Room Type are different from the original.

DONE:
(1) Found the two lines of code
 in 2_city_specific_clean.do responsible for reducing Full Data observations: sex_res > 2 & race_res > 4
(2) Replaced these with a command to replace sample = 0 instead of dropping, which fixed the discrepancy in the Full Data our tables were generated and the original tables.
(3) Changed variable header in Summary Table 1 from "Property Types" to "Property Type"