Table 4 is done. Identified Total OBS number issue. Other issues arose.
(1) Total Number of obsesrvations issue: Now total observation number = 17 050.  Issue was the drop command in  review_cleaner.do

(2.a) Changes made so that the reviewers.do would run properly: removed lines 66-71 and pasted into review_cleaner.do
(2.a) I have a concern with definitions of unkown. "Unkown" in rev_race_res for example is coded as the union of 'unknown' and other irrelevant categoreis and does not contain 'missings'. I edited to the code to highlight this issue

(3) 01_summary_table and 03_summary_host_characteristics had small bugs in latex portion which were fixed

(4) 04_summary_review_sentiment was edited to mimic table 4 in the paper.
IMPORTANT: I am concerned that the observations row is misleading in table 4. Host race and Reviewer race are different categorical variables with different numbers of people in each cell. The obersvations row only tells us the number of observations in row 1.

(5) Document tex was edited slightly.