Addressing email
(1)  What’s going on in the notes of Table 1? It says “99% of listings have strict (43%) cancelation policies” 

Table 1: Original sentence: "Only the most popular cancellation policy type is listed - in the full sample, 99\% of listings have strict (43\%), flexible (31\%) or moderate (25\%) cancellation policies."
What was meant by that phrase was that 99% of the full data can be broken down into 3 categories "strict", "flexible", "moderate". [The remaining 1% is a collection of 3 other categories].

I rewrote it as: "Strict cancellation policy is the most common out of 4 possible policies: strict (43\%), flexible (31\%), moderate (25\%) and other (1\%)"


(2) Add “Summary polarity” and “Summary subjectivity” as variables to Table 3,  after short words in summary.

I wasn't entirely sure how to do this. I made sure that all of the files are in the correct place and are being run when called upon. Now, when we run the run_project we also have all of the sentiment analysis data. That said, I wasn't sure what to add. Is it an average of all of the polarity and subjectivity categories? If so, how should this average be taken? Many of the observations which are marked as 0 were simply missing values rather than neutral sentences. description and summary are the two categories with most data. Idea: we could also do NLP on "name". "name" has the most observations, is the first thing people see on airbnb and often includes positive words. Food for thought. 

(3)  Table 4 - in title, do we mean reviewer race "in All data"? or is it actually “in Chicago data”? 
I think Chicago data is more appropriate. Should we also change "Full data" to  "Full Chicago Data"

(4) Effect on listing availability table: observations = 356 is super low. What happened? Do we have missing dummies for the census data? i bet it's dropping all those observations because a lot of census controls are missing. 
Similar problem in Table 12 - why does # observations drop down to 356 in Model 4? 
NOT FIXED.
I commented out the reviewers bit in run_project and ran project. 
I naively ran the regression in 8_availibility_30.do and checked the number of observations. There were roughly 508.
I then checked the code and there seemed to be a couple of odd errors in the 2_city_specific_clean folder. For one, full_listings_all_sentiments_updated6.dta  was not in the folders, so I uploaded it in my latest commit. That is now working. I reran 8_availibility_30 and still got 508. I wasn't able to find the issue and called it a day. 


I also added: save "$repository/data/cleaned.dta", replace in the run_project, so that we do not have to always rerun the time consuming code when we want to switch from reviewers analysis and main analysis. 

Additional concern: 
I am not sure whether we are accounting for all of the missing variables in all of the regression. Will check tomorrow.